align processing block - CUDA implementation added
