+++ src/rs.cpp
@@ -17,7 +17,7 @@
 #include "source.h"
 #include "core/processing.h"
 #include "proc/synthetic-stream.h"
-#include "proc/align.h"
+#include "proc/processing-blocks-factory.h"
 #include "proc/colorizer.h"
 #include "proc/pointcloud.h"
 #include "proc/disparity-transform.h"
@@ -1759,7 +1759,8 @@ rs2_processing_block* rs2_create_align(rs2_stream align_to, rs2_error** error) B
 {
     VALIDATE_ENUM(align_to);
 
-    auto block = std::make_shared<librealsense::align>(align_to);
+    auto block = create_align(align_to);
+
     return new rs2_processing_block{ block };
 }
 HANDLE_EXCEPTIONS_AND_RETURN(nullptr, align_to)
